refine merge operation

+++ src/core/common/ops/mmap_ops.rs
@@ -8,7 +8,7 @@ use super::madvise;
 
 pub const TEMP_FILE_EXTENSION: &str = "tmp";
 
-pub fn create_and_ensure_length(path: &Path, length: usize) -> Result<File, io::Error> {
+pub fn create_and_ensure_length(path: &Path, length: u64) -> Result<File, io::Error> {
     if path.exists() {
         let file = OpenOptions::new()
             .read(true)
@@ -17,7 +17,7 @@ pub fn create_and_ensure_length(path: &Path, length: usize) -> Result<File, io::
             // Don't truncate because we explicitly set the length later
             .truncate(false)
             .open(path)?;
-        file.set_len(length as u64)?;
+        file.set_len(length)?;
 
         Ok(file)
     } else {
@@ -32,7 +32,7 @@ pub fn create_and_ensure_length(path: &Path, length: usize) -> Result<File, io::
                 // Don't truncate because we explicitly set the length later
                 .truncate(false)
                 .open(&temp_path)?;
-            temp_file.set_len(length as u64)?;
+            temp_file.set_len(length)?;
         }
 
         std::fs::rename(&temp_path, path)?;
